requirements.txt
langchain
streamlit
python-dotenv
google-generativeai
st-chat-message
pandas
youtube-transcript-api
beautifulsoup4
tiktoken
